test: one file per router; test_fidelity.py dissolves into them
test_fidelity.py is gone. Its 37 tests were named for a campaign ("the new BYO fields +
correctness fixes") and sliced across ten providers, so a provider's shape assertions lived in two
files. Whether a test drives HTTP or calls a response builder directly is how it reaches the
assertion, not a layer — both assert "this provider's response has this shape" — so both go in the
provider's file.

New: test_github/hubspot/slack/atlassian/notion/s3/fireflies.py, joining test_google.py and
test_linear.py. test_endpoints.py -> test_cross_vendor.py (130 lines): the assertions that span
vendors and so belong to no single router.

The per-vendor crawlers move to _helpers — both each provider's "an admin crawl reaches every
document" and the cross-vendor subset check need them. test_fidelity's generic helpers become
_helpers.tiny_corpus / corpus_client / bare_request / epoch_of.

Dependency-checked before moving (0 cross-provider references after fixing two false positives in
my classifier). Verified by comparing the set of test function NAMES against the previous commit:
787 before, 787 after, none lost, none gained — which is how I caught that the first attempt
silently dropped 8 tests whose provider fell outside the seven new buckets.

pytest 860 passed; test_sdk + test_mcp + test_llamaindex 28 passed.
