+++ tests/_helpers.py
@@ -86,3 +86,164 @@ def db_count(conn, source_type, **kw) -> int:
 def tok(tokens_yaml, email: str) -> str:
     """One user's bearer token out of ``tokens.yaml``."""
     return next(u["token"] for u in tokens_yaml["users"] if u["email"] == email)
+
+
+def tiny_corpus(tmp_path, records):
+    """One small corpus in ``tmp_path``. Many shape tests call a router's response builder against
+    the resulting rows rather than over HTTP, so they want the settings, not a client."""
+    return build_corpus(tmp_path, records, name="corpus.jsonl")
+
+
+@contextlib.contextmanager
+def corpus_client(tmp_path, records):
+    """``records`` built and served, yielding ``(client, settings)`` — the GraphQL tests need the
+    admin token off the settings alongside the client."""
+    settings = tiny_corpus(tmp_path, records)
+    with client_for(settings) as client:
+        yield client, settings
+
+
+def bare_request():
+    """A minimal Starlette Request, for response builders that only read the URL."""
+    from starlette.requests import Request
+
+    return Request({"type": "http", "headers": [], "query_string": b"", "scheme": "http",
+                    "server": ("mock", 80), "path": "/"})
+
+
+def epoch_of(iso: str) -> int:
+    """An ISO-8601 string as unix seconds."""
+    from datetime import datetime
+
+    return int(datetime.fromisoformat(iso.replace("Z", "+00:00")).timestamp())
+
+
+# --- per-vendor crawlers -----------------------------------------------------------------------
+# Small page sizes on purpose, so each one exercises its vendor's pagination. Shared because two
+# kinds of test need them: each provider's "an admin crawl reaches every stored document", and the
+# cross-cutting check that a non-admin's crawl is a strict subset of the admin's.
+
+
+
+def crawl_slack(client, headers):
+    total, cursor = 0, None
+    channels = []
+    while True:
+        data = {"limit": 8}
+        if cursor:
+            data["cursor"] = cursor
+        j = client.post("/slack/api/conversations.list", headers=headers, data=data).json()
+        channels += j["channels"]
+        cursor = j["response_metadata"]["next_cursor"]
+        if not cursor:
+            break
+    for ch in channels:
+        ccur = None
+        while True:
+            d = {"channel": ch["id"], "limit": 50}
+            if ccur:
+                d["cursor"] = ccur
+            h = client.post("/slack/api/conversations.history", headers=headers, data=d).json()
+            for m in h["messages"]:
+                total += 1
+                if m.get("reply_count"):  # a thread root — its replies come from conversations.replies
+                    r = client.post("/slack/api/conversations.replies", headers=headers,
+                                    data={"channel": ch["id"], "ts": m["ts"]}).json()
+                    total += len(r["messages"]) - 1  # thread includes the root we already counted
+            ccur = h["response_metadata"]["next_cursor"]
+            if not ccur:
+                break
+    return total
+
+
+def crawl_hubspot(client, headers, object_type, limit=2, archived=False):
+    """Cursor-paginate one CRM object type. Terminates on the ABSENCE of paging.next — which is
+    exactly how the official client's fetch_all decides it is done, so a mock that always emits
+    paging.next would hang a real client rather than error."""
+    out, after = [], None
+    while True:
+        params = {"limit": limit}
+        if archived:
+            params["archived"] = "true"
+        if after:
+            params["after"] = after
+        j = client.get(f"/hubspot/crm/v3/objects/{object_type}", headers=headers,
+                       params=params).json()
+        out += j["results"]
+        nxt = (j.get("paging") or {}).get("next")
+        if not nxt:
+            break
+        after = nxt["after"]
+    return out
+
+
+
+# --- crawlers (small page sizes to exercise pagination) -------------------------
+
+def crawl_gmail(client, headers, user="me"):
+    ids, token = [], None
+    while True:
+        p = {"maxResults": 7}
+        if token:
+            p["pageToken"] = token
+        j = client.get(f"/gmail/v1/users/{user}/messages", headers=headers, params=p).json()
+        ids += [m["id"] for m in j.get("messages", [])]
+        token = j.get("nextPageToken")
+        if not token:
+            break
+    return ids
+
+
+def crawl_drive(client, headers):
+    ids, token = [], None
+    while True:
+        p = {"pageSize": 7}
+        if token:
+            p["pageToken"] = token
+        j = client.get("/drive/v3/files", headers=headers, params=p).json()
+        ids += [f["id"] for f in j.get("files", [])]
+        token = j.get("nextPageToken")
+        if not token:
+            break
+    return ids
+
+
+
+def crawl_github_repo(client, headers, org, repo):
+    out, page = [], 1
+    while True:
+        r = client.get(f"/github/repos/{org}/{repo}/issues", headers=headers,
+                       params={"per_page": 5, "page": page, "state": "all"})
+        body = r.json()
+        out += body
+        if 'rel="next"' not in r.headers.get("Link", ""):
+            break
+        page += 1
+    return out
+
+
+
+def crawl_jira(client, headers):
+    out, token = [], None
+    while True:
+        p = {"maxResults": 6}
+        if token:
+            p["nextPageToken"] = token
+        j = client.get("/atlassian/rest/api/3/search/jql", headers=headers, params=p).json()
+        out += j["issues"]
+        if j.get("isLast", True):
+            break
+        token = j["nextPageToken"]
+    return out
+
+
+def crawl_confluence(client, headers):
+    out, start, limit = [], 0, 7
+    while True:
+        j = client.get("/atlassian/wiki/rest/api/content", headers=headers,
+                       params={"start": start, "limit": limit, "expand": "body.storage"}).json()
+        out += j["results"]
+        if "next" not in j.get("_links", {}):
+            break
+        start += limit
+    return out
